refactor: remove App from simple templates
